.server-fastapi/requirements.txt
fastapi[standard]
jinja2
sqlalchemy
mysql-connector-python
uvicorn